root/list_models.py
from google import genai
import os
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("gemini_api_key"))

models = client.models.list()

for m in models:
    print(m.name)